requirements.txt
groq==0.25.0
pyTelegramBotAPI==4.27.0
requests==2.32.3
